+++ draft-packages/modal/KaizenDraft/Modal/Presets/InputEditModal.tsx
@@ -1,7 +1,6 @@
+import classnames from "classnames"
 import * as React from "react"
-
-import { Text } from "@kaizen/component-library"
-
+import { Heading } from "@kaizen/component-library"
 import { ButtonProps } from "@kaizen/draft-button"
 import {
   GenericModal,
@@ -10,12 +9,12 @@ import {
   ModalFooter,
   ModalHeader,
 } from "../"
-
 import styles from "./InputEditModal.scss"
 
 export interface InputEditModalProps {
   readonly isOpen: boolean
-  readonly type: "positive" | "negative"
+  readonly unpadded?: boolean
+  readonly mood: "positive" | "destructive"
   readonly title: string
   readonly onSubmit: () => void
   readonly onDismiss: () => void
@@ -31,7 +30,7 @@ type InputEditModal = React.FunctionComponent<InputEditModalProps>
 
 const InputEditModal = ({
   isOpen,
-  type,
+  mood,
   title,
   onSubmit,
   localeDirection = "ltr",
@@ -40,6 +39,7 @@ const InputEditModal = ({
   submitWorking,
   automationId,
   children,
+  unpadded = false,
   ...props
 }: InputEditModalProps) => {
   const onDismiss = submitWorking ? undefined : props.onDismiss
@@ -64,24 +64,34 @@ const InputEditModal = ({
       automationId={automationId}
     >
       <div className={styles.modal} dir={localeDirection}>
-        <ModalHeader unpadded onDismiss={onDismiss}>
-          <div className={styles.header}>
+        <ModalHeader onDismiss={onDismiss}>
+          <div
+            className={classnames(styles.header, {
+              [styles.textAlignRTL]: localeDirection === "rtl",
+              [styles.padded]: !unpadded,
+            })}
+          >
             <ModalAccessibleLabel>
-              <Text tag="h1" style="zen-heading-3" inline>
+              <Heading tag="h2" variant="heading-2">
                 {title}
-              </Text>
+              </Heading>
             </ModalAccessibleLabel>
           </div>
         </ModalHeader>
-        <ModalBody unpadded>
-          <div className={styles.body} dir={localeDirection}>
+        <ModalBody>
+          <div
+            className={classnames(styles.body, { [styles.padded]: !unpadded })}
+            dir={localeDirection}
+          >
             {children}
           </div>
         </ModalBody>
         <ModalFooter
           actions={footerActions}
-          appearance={type === "negative" ? "destructive" : "primary"}
+          appearance={mood === "destructive" ? "destructive" : "primary"}
           automationId={automationId}
+          variant="inputEdit"
+          unpadded={unpadded}
         />
       </div>
     </GenericModal>