+++ packages/notification/docs/ToastNotification.stories.tsx
@@ -75,7 +75,7 @@ export default {
       },
     },
     ...figmaEmbed(
-      "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=1929%3A21830"
+      "https://www.figma.com/file/eZKEE5kXbEMY3lx84oz8iN/%E2%9D%A4%EF%B8%8F-UI-Kit%3A-Heart?node-id=1929%3A21830"
     ),
   },
   decorators: [withDesign, withNavigation],
@@ -84,7 +84,7 @@ export default {
 export const PositiveKaizenSiteDemo = () => {
   React.useEffect(() => {
     addToastNotification({
-      type: "affirmative",
+      type: "positive",
       title: "Success",
       automationId: "notification1",
       message: (
@@ -104,7 +104,7 @@ export const PositiveAutohide = () => (
   <Triggers
     notifications={[
       {
-        type: "affirmative",
+        type: "positive",
         title: "Success",
         automationId: "notification1",
         message: (
@@ -125,7 +125,7 @@ export const PositiveAutohideHideCloseIcon = () => (
     notifications={[
       {
         automationId: "notification1",
-        type: "affirmative",
+        type: "positive",
         title: "Success",
         message: (
           <>
@@ -199,7 +199,7 @@ export const MultipleNotifications = () => (
     notifications={[
       {
         automationId: "notification1",
-        type: "affirmative",
+        type: "positive",
         title: "Success",
         message: (
           <>
@@ -253,7 +253,7 @@ export const OverflowNotifications = () => {
     <Triggers
       notifications={[...Array(15)].map((_, i) => ({
         automationId: `notification-${i}-${seed}`,
-        type: Math.random() > 0.5 ? "affirmative" : "cautionary",
+        type: Math.random() > 0.5 ? "positive" : "cautionary",
         title: "Success",
         autohide: false,
         message: (
@@ -279,7 +279,7 @@ export const UpdatedNotification = () => (
               addToastNotification({
                 id: "consistent-id",
                 automationId: "notification1",
-                type: "affirmative",
+                type: "positive",
                 title: "Initial notification title",
                 message: (
                   <>