+++ draft-packages/likert-scale-legacy/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.26](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-likert-scale-legacy@1.3.25...@kaizen/draft-likert-scale-legacy@1.3.26) (2021-12-20)
+
+**Note:** Version bump only for package @kaizen/draft-likert-scale-legacy
+
+
+
+
+
+## [1.3.25](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-likert-scale-legacy@1.3.24...@kaizen/draft-likert-scale-legacy@1.3.25) (2021-12-13)
+
+**Note:** Version bump only for package @kaizen/draft-likert-scale-legacy
+
+
+
+
+
+## [1.3.24](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-likert-scale-legacy@1.3.23...@kaizen/draft-likert-scale-legacy@1.3.24) (2021-12-08)
+
+**Note:** Version bump only for package @kaizen/draft-likert-scale-legacy
+
+
+
+
+
 ## [1.3.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-likert-scale-legacy@1.3.22...@kaizen/draft-likert-scale-legacy@1.3.23) (2021-11-15)
 
 **Note:** Version bump only for package @kaizen/draft-likert-scale-legacy