+++ packages/tabs/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.1.12...@kaizen/tabs@1.1.13) (2021-12-13)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
+## [1.1.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.1.11...@kaizen/tabs@1.1.12) (2021-12-13)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
+## [1.1.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.1.10...@kaizen/tabs@1.1.11) (2021-12-09)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
+## [1.1.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.1.9...@kaizen/tabs@1.1.10) (2021-12-08)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
 ## [1.1.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.1.8...@kaizen/tabs@1.1.9) (2021-12-02)
 
 **Note:** Version bump only for package @kaizen/tabs