+++ site/CHANGELOG.md
@@ -3,6 +3,110 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.0.0-private.40](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.39...site@0.0.0-private.40) (2021-12-24)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.39](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.38...site@0.0.0-private.39) (2021-12-22)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.38](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.37...site@0.0.0-private.38) (2021-12-20)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.37](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.36...site@0.0.0-private.37) (2021-12-20)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.36](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.35...site@0.0.0-private.36) (2021-12-17)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.35](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.34...site@0.0.0-private.35) (2021-12-15)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.34](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.33...site@0.0.0-private.34) (2021-12-13)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.33](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.32...site@0.0.0-private.33) (2021-12-13)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.32](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.31...site@0.0.0-private.32) (2021-12-12)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.31](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.30...site@0.0.0-private.31) (2021-12-09)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.30](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.29...site@0.0.0-private.30) (2021-12-08)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.29](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.28...site@0.0.0-private.29) (2021-12-06)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.28](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.27...site@0.0.0-private.28) (2021-12-06)
+
+**Note:** Version bump only for package site
+
+
+
+
+
 # [0.0.0-private.27](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.26...site@0.0.0-private.27) (2021-12-02)
 
 **Note:** Version bump only for package site